+++ src/main/java/com/ycloud/client/model/WhatsappMessageInteractiveAction.java
@@ -91,13 +91,15 @@ public class WhatsappMessageInteractiveAction {
   private List<WhatsappMessageInteractiveActionSection> sections = null;
 
   /**
-   * Action name. Required for Call-To-Action (CTA) URL Button Messages. - &#x60;cta_url&#x60;: Use for Call-To-Action (CTA) URL Buttons. - &#x60;send_location&#x60;: Use for Location Request Messages.
+   * Action name. Required for Call-To-Action (CTA) buttons. - &#x60;cta_url&#x60;: Use for Call-To-Action (CTA) URL buttons. - &#x60;send_location&#x60;: Use for Location Request buttons. - &#x60;flow&#x60;: Use for Flow buttons.
    */
   @JsonAdapter(NameEnum.Adapter.class)
   public enum NameEnum {
     CTA_URL("cta_url"),
     
-    SEND_LOCATION("send_location");
+    SEND_LOCATION("send_location"),
+    
+    FLOW("flow");
 
     private String value;
 
@@ -139,7 +141,7 @@ public class WhatsappMessageInteractiveAction {
 
   public static final String SERIALIZED_NAME_NAME = "name";
   /**
-   * Action name. Required for Call-To-Action (CTA) URL Button Messages. - &#x60;cta_url&#x60;: Use for Call-To-Action (CTA) URL Buttons. - &#x60;send_location&#x60;: Use for Location Request Messages.
+   * Action name. Required for Call-To-Action (CTA) buttons. - &#x60;cta_url&#x60;: Use for Call-To-Action (CTA) URL buttons. - &#x60;send_location&#x60;: Use for Location Request buttons. - &#x60;flow&#x60;: Use for Flow buttons.
   **/
   @SerializedName(SERIALIZED_NAME_NAME)
   private NameEnum name;
@@ -306,7 +308,7 @@ public class WhatsappMessageInteractiveAction {
   }
 
    /**
-   * Action name. Required for Call-To-Action (CTA) URL Button Messages. - &#x60;cta_url&#x60;: Use for Call-To-Action (CTA) URL Buttons. - &#x60;send_location&#x60;: Use for Location Request Messages.
+   * Action name. Required for Call-To-Action (CTA) buttons. - &#x60;cta_url&#x60;: Use for Call-To-Action (CTA) URL buttons. - &#x60;send_location&#x60;: Use for Location Request buttons. - &#x60;flow&#x60;: Use for Flow buttons.
    * @return name
   **/
   @javax.annotation.Nullable
@@ -317,9 +319,9 @@ public class WhatsappMessageInteractiveAction {
 
 
   /**
-   * Action name. Required for Call-To-Action (CTA) URL Button Messages. - &#x60;cta_url&#x60;: Use for Call-To-Action (CTA) URL Buttons. - &#x60;send_location&#x60;: Use for Location Request Messages.
+   * Action name. Required for Call-To-Action (CTA) buttons. - &#x60;cta_url&#x60;: Use for Call-To-Action (CTA) URL buttons. - &#x60;send_location&#x60;: Use for Location Request buttons. - &#x60;flow&#x60;: Use for Flow buttons.
    **/
-  @ApiModelProperty(value = "Action name. Required for Call-To-Action (CTA) URL Button Messages. - `cta_url`: Use for Call-To-Action (CTA) URL Buttons. - `send_location`: Use for Location Request Messages.")
+  @ApiModelProperty(value = "Action name. Required for Call-To-Action (CTA) buttons. - `cta_url`: Use for Call-To-Action (CTA) URL buttons. - `send_location`: Use for Location Request buttons. - `flow`: Use for Flow buttons.")
   public void setName(NameEnum name) {
     this.name = name;
   }